Todo el host compila, aún explota

+++ src/SourceFiles/ControllerGameMode.cpp
@@ -34,7 +34,7 @@ void ControllerGameMode::render() {
 	if (roundFinished_) {
 		Texture* ganador = SDL_Game::instance()->getTexturesMngr()->getTexture(Resources::winner1 + winnerId_);
 		SDL_Rect destRect{
-			halfWidth_ - ganador->getWidth() * 0.5f,
+			halfWidth_ - (int)(ganador->getWidth() * 0.5),
 			halfHeight_,
 			ganador->getWidth(),
 			ganador->getHeight()
