+++ src/SourceFiles/WiFightGameMode.cpp
@@ -32,7 +32,7 @@ void WiFightGameMode::render()
 	if (roundFinished_) {
 		Texture* ganador = SDL_Game::instance()->getTexturesMngr()->getTexture(Resources::winner1 + winnerId_);
 		SDL_Rect destRect{
-			halfWinWidth_ - ganador->getWidth() * 0.5f,
+			halfWinWidth_ - (int)(ganador->getWidth() * 0.5f),
 			halfWinHeight_,
 			ganador->getWidth(),
 			ganador->getHeight()
